backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
google-generativeai==0.8.0
python-multipart==0.0.6
pydantic==2.5.3
email-validator==2.1.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.13.1
bcrypt==3.2.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
requests==2.31.0
